pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "anomaly_detection"
version = "0.0.1"
description = "ML project for the 02476 MLOPS course at DTU. Using DINOv3 to find anomalies in object from images."
authors = [
  { name = "Team15", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bentoml>=1.4.33",
    "fastapi==0.115.6",
    "google-api-python-client>=2.187.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "locust>=2.43.1",
    "markdown>=3.10",
    "matplotlib>=3.10.8",
    "mkdocs>=1.6.1",
    "mkdocs-material>=4.8.0",
    "mkdocstrings[python]>=0.28.2",
    "numpy>=2.0,<2.4",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.7.2",
    "snakeviz>=2.2.2",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
]

[project.optional-dependencies]
frontend = [
  "streamlit>=1.30",
  "requests>=2.31",
  "pillow>=10.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.ruff.lint.per-file-ignores]
"src/anomaly_detection/*_hydra.py" = ["E402"]
"src/anomaly_detection/api.py" = ["E402"]
"src/anomaly_detection/api_inference.py" = ["E402"]
"src/anomaly_detection/evaluate.py" = ["E402"]
"src/anomaly_detection/inference.py" = ["E402"]
"src/anomaly_detection/model.py" = ["E402"]
"src/anomaly_detection/train.py" = ["E402"]
"src/anomaly_detection/vision_transformer.py" = ["E402"]
"src/anomaly_detection/layers/__init__.py" = ["F401"]

[tool.coverage.run]
source = ["src/anomaly_detection"]
omit = ["tests/*" , "__init__.py" , "src/anomaly_detection/*hydra*.py", "src/anomaly_detection/layers/*", "src/anomaly_detection/utils.py", "src/anomaly_detection/vision_transformer.py"]
